Add a mime power implanter
