+++ Content.Shared/_NF/Implants/MimePowersImplantComponent.cs
@@ -3,7 +3,7 @@ using Robust.Shared.GameStates;
 namespace Content.Shared._NF.Implants.Components;
 
 /// <summary>
-/// Implant to get BibleUser status (to pray, summon familiars, bless with bibles)
+/// Implant to get MimePowers status (to summon walls and take the mime's vow)
 /// </summary>
 [RegisterComponent, NetworkedComponent]
-public sealed partial class BibleUserImplantComponent : Component;
+public sealed partial class MimePowersImplantComponent : Component;
