+++ Resources/Prototypes/_NF/Entities/Objects/Misc/implanters.yml
@@ -39,6 +39,14 @@
   - type: Implanter
     implant: BibleUserImplant
 
+- type: entity
+  id: MimePowersImplanter
+  suffix: faith
+  parent: BaseImplantOnlyImplanter
+  components:
+  - type: Implanter
+    implant: MimePowersImplant
+
 # Labeled implanters (for vendors)
 - type: entity
   id: MedicalTrackingImplanterLabeled
