+++ Resources/Prototypes/_NF/Entities/Objects/Misc/subdermal_implants.yml
@@ -110,6 +110,17 @@
 
 - type: entity
   parent: BaseSubdermalImplant
+  id: MimePowersImplant
+  name: mime power implant
+  description: This implant lets the user take the mime's vow.
+  categories: [ HideSpawnMenu ]
+  components:
+  - type: SubdermalImplant
+    permanent: true
+  - type: MimePowersImplant
+
+- type: entity
+  parent: BaseSubdermalImplant
   id: RedLightImplant
   name: light implant (red)
   description: This implant emits light from the user's skin on activation.
