+++ Resources/Prototypes/_NF/Loadouts/contractor_loadout_groups.yml
@@ -880,6 +880,7 @@
   - ContractorLightImplanter
   - ContractorBikeHornImplanter
   - ContractorSadTromboneImplanter
+  - ContractorMimeImplanter
   fallbacks:
   - ContractorMedicalTrackingImplanter
