+++ Resources/Prototypes/_NF/Mail/mail.yml
@@ -1912,6 +1912,7 @@
         orGroup: Mask
       - id: ClothingMaskScaredMime
         orGroup: Mask
+      - id: MimePowersImplanter
 
 # Base Nyano mail
